+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🚨 Fixes
+- Update `ICanRegenerateGithubWorkflows` component to use `fallout` tool instead of `nuke`
 
 ## [3.0.0] / 2026-06-19
 ### 💥 Breaking changes